Module substrate_lite::chain::sync::full_optimistic

Optimistic header and body syncing.

This state machine builds, from a set of sources, a fully verified chain of blocks headers and bodies.

In addition to managing the sources, using OptimisticFullSync also requires holding the storage of the latest finalized block.

Re-exports

pub use optimistic::FinishRequestOutcome;

pub use optimistic::RequestAction;

pub use optimistic::RequestFail;

pub use optimistic::RequestId;

pub use optimistic::SourceId;

pub use optimistic::Start;

Structs

Block

Config
Configuration for the OptimisticFullSync.

InProgress
Resume the processing after a ProcessOne::InProgress.

OptimisticFullSync
Optimistic headers-only syncing.

RequestSuccessBlock

StorageGet
Loading a storage value is required in order to continue.

StorageNextKey
Fetching the key that follows a given one is required in order to continue.

StoragePrefixKeys
Fetching the list of keys with a given prefix is required in order to continue.

Enums

ProcessOne
State of the processing of blocks.