Module substrate_lite::chain::sync::headers_optimistic

Optimistic headers-only syncing.

Optimistic syncing consists in assuming that all sources of blocks form the same chain. A query for blocks is performed on a random source, and the response is verified. If it turns out that the source doesn't belong to the same chain (or is malicious), a different source is tried.

While this syncing strategy is very simplistic, it is the most effective when the majority of sources are well-behaved, which is normally the case.

The OptimisticHeadersSync makes it possible to sync the finalized blocks of a chain, but not the non-finalized blocks.

Re-exports

pub use optimistic::FinishRequestOutcome;

pub use optimistic::RequestAction;

pub use optimistic::RequestFail;

pub use optimistic::RequestId;

pub use optimistic::SourceId;

pub use optimistic::Start;

Structs

Config
Configuration for the OptimisticHeadersSync.

OptimisticHeadersSync
Optimistic headers-only syncing.

RequestSuccessBlock
Single block in the outcome of a request. A list of these must be passed to OptimisticHeadersSync::finish_request.

Enums

ProcessOneOutcome
Outcome of calling OptimisticHeadersSync::process_one.

ResetCause
Problem that happened and caused the reset.