Module substrate_lite::chain::sync::optimistic

Code shared in common between headers-only and headers-and-body optimistic syncing.

Structs

Config
Configuration for the OptimisticSync.

OptimisticSync
Optimistic headers-only syncing.

ProcessOne

ProcessOneReport

RequestId
Identifier for an ongoing request in the OptimisticSync.

RequestsDrain
Iterator that drains requests after a source has been removed.

SourceId
Identifier for a source in the OptimisticSync.

Start
Must be used to accept the request.

Enums

FinishRequestOutcome

RequestAction
Request that should be emitted towards a certain source.

RequestFail
Reason why a request has failed.