Module substrate_lite::network

Re-exports

pub use peer_id::PeerId;

Modules

connection
Module containing everything related to the processing of a single libp2p connection.

discovery
Mechanisms related to discovering nodes that are part of a certain overlay network.

leb128
Little Endian Base 128

libp2p
Collection of libp2p connections.

multiaddr

peer_id

peerset
Data structure storing a networking state. Helper for managing connectivity to overlay networks.

protocol
Substrate/Polkadot-specific protocols on top of libp2p.

service

Structs

Multiaddr
Representation of a Multiaddr.