Module substrate_lite::network::libp2p

Collection of libp2p connections.

The Network struct in this module is a collection of libp2p connections. It uses internal buffering and interior mutability in order to provide a convenient-to-use API based around notifications protocols and request-response protocols.

Re-exports

pub use connection::established::ConfigRequestResponse;

Structs

Config
Configuration for a Network.

ConnectionId
Identifier of a connection spawned by the Network.

Network
Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

OverlayNetwork
Configuration for a specific overlay network.

PendingId
Identifier of a pending connection requested by the network through a StartConnect.

ReadWrite
Outcome of calling Network::read_write.

StartConnect
User must start connecting to the given multiaddress.

SubstreamOpen

Enums

ConnectionError

Event
Event generated by Network::next_event.

RequestError