Module substrate_lite::network::protocol

Substrate/Polkadot-specific protocols on top of libp2p.

Libp2p provides a framework on top of which higher-level, domain-specific protocols can be used. This module contains the domain-specific protocols specific to Substrate and Polkadot.

Structs

BlockAnnounceRef
Decoded block announcement notification.

BlockAnnouncesHandshakeDecodeError
Error potentially returned by decode_block_announces_handshake.

BlockAnnouncesHandshakeRef
Decoded handshake sent or received when opening a block announces notifications substream.

BlockData
Block sent in a block response.

BlocksRequestConfig
Description of a block request that can be sent to a peer.

BlocksRequestFields
Which fields should be present in the response.

DecodeBlockAnnounceError
Error potentially returned by decode_block_announces_handshake.

ProtobufDecodeError
Error while decoding the protobuf encoding.

StorageProofRequestConfig
Description of a storate proof request that can be sent to a peer.

Enums

BlocksRequestConfigStart
Which block the remote must return first.

BlocksRequestDirection
Whether the first block should be the one with the highest number, of the one with the lowest number.

DecodeBlockResponseError
Error potentially returned by decode_block_response.

DecodeStorageProofResponseError
Error potentially returned by decode_storage_proof_response.

Role
Role a node reports playing on the network.

Functions

build_block_request
Builds the bytes corresponding to a block request.

build_storage_proof_request
Builds the bytes corresponding to a storage proof request.

decode_block_announce
Decodes a block announcement.

decode_block_announces_handshake
Decodes a SCALE-encoded block announces handshake.

decode_block_response
Decodes a response to a block request.

decode_storage_proof_response
Decodes a response to a storage proof request.

encode_block_announce
Turns a block announcement into its SCALE-encoding ready to be sent over the wire.

encode_block_announces_handshake
Turns a block announces handshake into its SCALE-encoding ready to be sent over the wire.